first attempt at recovery reorganisation
