+++ examples/compressible/dry_bryan_fritsch.py
@@ -106,17 +106,17 @@ Vu_DG1 = VectorFunctionSpace(mesh, VDG1.ufl_element())
 Vu_CG1 = VectorFunctionSpace(mesh, "CG", 1)
 Vu_brok = FunctionSpace(mesh, BrokenElement(Vu.ufl_element()))
 
-u_opts = RecoveredOptions(embedding_space=Vu_DG1,
-                          recovered_space=Vu_CG1,
-                          broken_space=Vu_brok,
-                          boundary_method=Boundary_Method.dynamics)
-rho_opts = RecoveredOptions(embedding_space=VDG1,
-                            recovered_space=VCG1,
-                            broken_space=Vr,
-                            boundary_method=Boundary_Method.dynamics)
-theta_opts = RecoveredOptions(embedding_space=VDG1,
-                              recovered_space=VCG1,
-                              broken_space=Vt_brok)
+u_opts = RecoveryOptions(embedding_space=Vu_DG1,
+                         recovered_space=Vu_CG1,
+                         broken_space=Vu_brok,
+                         boundary_method=Boundary_Method.dynamics)
+rho_opts = RecoveryOptions(embedding_space=VDG1,
+                           recovered_space=VCG1,
+                           broken_space=Vr,
+                           boundary_method=Boundary_Method.dynamics)
+theta_opts = RecoveryOptions(embedding_space=VDG1,
+                             recovered_space=VCG1,
+                             broken_space=Vt_brok)
 
 transported_fields = [SSPRK3(state, "rho", options=rho_opts),
                       SSPRK3(state, "theta", options=theta_opts),
