+++ gusto/recovery/__init__.py
@@ -1,0 +1,4 @@
+from gusto.recovery.averaging import *            # noqa
+from gusto.recovery.recovery import *             # noqa
+from gusto.recovery.recovery_kernels import *     # noqa
+from gusto.recovery.reversible_recovery import *  # noqa
