+++ integration-tests/transport/test_recovered_transport.py
@@ -37,12 +37,12 @@ def test_recovered_space_setup(tmpdir, geometry, tracer_setup):
     state.fields("u").project(setup.uexpr)
 
     # Declare transport scheme
-    recovered_opts = RecoveredOptions(embedding_space=VDG1,
-                                      recovered_space=VCG1,
-                                      broken_space=VDG0,
-                                      boundary_method=Boundary_Method.dynamics)
+    recovery_opts = RecoveryOptions(embedding_space=VDG1,
+                                    recovered_space=VCG1,
+                                    broken_space=VDG0,
+                                    boundary_method=Boundary_Method.dynamics)
 
-    transport_scheme = [(eqn, SSPRK3(state, options=recovered_opts))]
+    transport_scheme = [(eqn, SSPRK3(state, options=recovery_opts))]
 
     # Run and check error
     error = run(state, transport_scheme, setup.tmax, setup.f_end)
